requirements.txt
requests
pillow
numpy
opencv-python
pytesseract
keyboard